+++ src/main/java/pl/thedeem/intellij/dql/completion/DQLLocalCompletionContributor.java
@@ -1,29 +1,13 @@
 package pl.thedeem.intellij.dql.completion;
 
 import com.intellij.codeInsight.completion.*;
-import com.intellij.patterns.PlatformPatterns;
 import com.intellij.psi.PsiElement;
 import com.intellij.util.ProcessingContext;
 import org.jetbrains.annotations.NotNull;
 import pl.thedeem.intellij.dql.completion.engines.*;
-import pl.thedeem.intellij.dql.psi.DQLTypes;
 
 public class DQLLocalCompletionContributor extends CompletionContributor {
     public DQLLocalCompletionContributor() {
-        extend(CompletionType.BASIC, DQLPsiPatterns.SUGGEST_FIELD_VALUES,
-                new CompletionProvider<>() {
-                    public void addCompletions(@NotNull CompletionParameters parameters, @NotNull ProcessingContext context, @NotNull CompletionResultSet resultSet) {
-                        new DQLExpressionsCompletions().autocomplete(parameters, resultSet);
-                    }
-                }
-        );
-        extend(CompletionType.BASIC, DQLPsiPatterns.SUGGEST_FIELD_VALUES,
-                new CompletionProvider<>() {
-                    public void addCompletions(@NotNull CompletionParameters parameters, @NotNull ProcessingContext context, @NotNull CompletionResultSet resultSet) {
-                        new DQLFieldValuesCompletions().autocomplete(parameters, resultSet);
-                    }
-                }
-        );
         extend(CompletionType.BASIC, DQLPsiPatterns.INSIDE_STATEMENT_PARAMETERS_LIST,
                 new CompletionProvider<>() {
                     public void addCompletions(@NotNull CompletionParameters parameters, @NotNull ProcessingContext context, @NotNull CompletionResultSet resultSet) {
@@ -66,6 +50,13 @@ public class DQLLocalCompletionContributor extends CompletionContributor {
                     }
                 }
         );
+        extend(CompletionType.BASIC, DQLPsiPatterns.INSIDE_PARAMETER,
+                new CompletionProvider<>() {
+                    public void addCompletions(@NotNull CompletionParameters parameters, @NotNull ProcessingContext context, @NotNull CompletionResultSet resultSet) {
+                        new DQLParameterValuesCompletion().autocomplete(parameters, resultSet);
+                    }
+                }
+        );
     }
 
     @Override
@@ -74,10 +65,7 @@ public class DQLLocalCompletionContributor extends CompletionContributor {
         PsiElement position = parameters.getPosition();
 
         // inside comments, we don't want any completions
-        if (PlatformPatterns.or(
-                PlatformPatterns.psiElement(DQLTypes.ML_COMMENT),
-                PlatformPatterns.psiElement(DQLTypes.EOL_COMMENT)
-        ).accepts(position)) {
+        if (DQLPsiPatterns.INSIDE_COMMENTS.accepts(position)) {
             return;
         }