+++ src/main/java/pl/thedeem/intellij/dql/psi/elements/impl/AbstractOperatorElementImpl.java
@@ -158,7 +158,8 @@ public abstract class AbstractOperatorElementImpl extends ASTWrapperPsiElement i
 
     private @Nullable Operator recalculateDefinition() {
         DQLDefinitionService service = DQLDefinitionService.getInstance(getProject());
-        return service.getOperator(getOperationId());
+        String id = getOperationId();
+        return service.getOperator(id);
     }
 
     private @Nullable Signature recalculateSignature() {
@@ -202,32 +203,14 @@ public abstract class AbstractOperatorElementImpl extends ASTWrapperPsiElement i
         return signature.outputs();
     }
 
-    protected String getOperationId() {
+    protected abstract String getOperationId();
+
+    @Override
+    public @Nullable String getOperatorSymbol() {
         ExpressionOperatorImpl operator = PsiTreeUtil.getChildOfType(this, ExpressionOperatorImpl.class);
         if (operator == null) {
             return null;
         }
-        return switch (operator.getText().trim().toLowerCase()) {
-            case "-" -> "dql.operator.subtract";
-            case "+" -> "dql.operator.add";
-            case "!=" -> "dql.operator.notEquals";
-            case "==" -> "dql.operator.equals";
-            case "and" -> "dql.operator.and";
-            case "not" -> "dql.operator.not";
-            case "or" -> "dql.operator.or";
-            case "xor" -> "dql.operator.xor";
-            case "/" -> "dql.operator.divide";
-            case ">" -> "dql.operator.greater";
-            case ">=" -> "dql.operator.greaterEquals";
-            case "<" -> "dql.operator.lower";
-            case "<=" -> "dql.operator.lowerEquals";
-            case "%" -> "dql.operator.modulo";
-            case "*" -> "dql.operator.multiply";
-            case "~" -> "dql.operator.search";
-            case "@" -> "dql.operator.timeAlignmentAt";
-            case "in" -> "dql.operator.in";
-            case "=" -> "dql.operator.assign";
-            default -> "";
-        };
+        return operator.getText().trim().toLowerCase();
     }
 }